Initial aim ahead development (#3)
* [ALS-5567] Update displayNotAuthorized function in login.js
The displayNotAuthorized function in login.js has been updated to include a dynamic login link to the notAuthorizedTemplate. This change improves maintainability and decreases potential for errors as URLs won't require hard coding anymore.

* [ALS-5566] Add showLoginPage function in login.js
A new function showLoginPage has been added to login.js that displays a confirmation message upon successful authentication. This is specifically implemented for BioDataCatalyst authentication and the message is shown during the UserProfile information processing phase.

* [GC] Remove unnecessary override files
* [GC] Add logo to images directory in Dockerfile
* [GC] Remove Open Access wording

* [GC] Update header links in interface
The href attribute in the interface header links has been updated to provide proper redirection functionality. This change corrects problems with dead links on the website, specifically in the header section, improving usability and the overall user experience.

* [ALS-5517] Remove unused functions in router.js
The commit removes displayDataAccess, getGenomicFilters, and displayAPI functions from router.js. These functions were not in use, contributing to complexity and potential confusion. The removal of these functions simplifies the codebase and makes maintenance easier.

* Remove Open PIC-SURE link from header
The Open PIC-SURE link in the header navigation has been removed as it was no longer needed. This change simplifies the header component and removes potential confusion for users navigating the UI.

* Add landing page and update header in PIC-SURE UI
* [ALS-5581] Remove footer content with override

* Added a function to generate random state for each authentication request in login.js to mitigate CSRF attacks. Adjusted behavior when a code is present and added checks for state parameter to further secure the login flow. Also improved error message handling for authentication failures.

* Add Super Admin and Admin navigation options
A Super Admin Console drop-down menu and an Admin Users button have been introduced in the header navigation. The Super Admin Console contains links for Application, Connection, Privilege, and Role management. These functionalities are only visible to users with respective Super Admin and Admin privileges.

* [ALS-5779] AIM-AHEAD PIC-SURE: Change header name from Authorized Access to Explore
* [ALS-5779] Redirect to login page if authentication was successful
* [ALS-5781] AIM-AHEAD PIC-SURE: Show the API tab
* [ALS-5800] Set logo as favicon
* [ALS-5781] AIM-AHEAD PIC-SURE: Show the API tab
* [ALS-5567] Fix order and now hide header-tabs during the login process

* Remove console log and inline redirect