[ALS-5476] Add header.js override file to fix TOS (#8)
* [ALS-5476] Add new header.js file for custom UI component
This commit adds a new header file in the pic-sure-aim-ahead-frontend project to override the default UI and add custom behavior after the main header has rendered. Specifically, it introduces logic to focus on the help dropdown menu when the terms of service modal is displayed.

* Refactor header.js to remove unnecessary import
In the header.js file, the "header/goToAuthorized" import was removed as it was unnecessary. By doing so, the code becomes cleaner and easier to maintain.