[ALS-5852] AIM-AHEAD PIC-SURE StatViz should work and not show obfuscation (#4)
* [ALS-5852] Update Wording

* [ALS-5852] Update footer render function

The update introduces new dependencies in the footer.js file and modifies the render function with additional functionality. A new middleware and redirect modal are instantiated in the function. An event handler is also added for clicks on anchors with target="_blank" to prevent the default behavior and render the redirect modal instead.