Minor UI fixes and updates (#9)
* [ALS-5476] Add new header.js file for custom UI component

This commit adds a new header file in the pic-sure-aim-ahead-frontend project to override the default UI and add custom behavior after the main header has rendered. Specifically, it introduces logic to focus on the help dropdown menu when the terms of service modal is displayed.

* [ALS-5699] Remove Icons and update sub search text

* [ALS-5850] Update GitHub Link

* [ALS-5850] Update logo title to say "AIM-AHEAD"

* [ALS-5852] Update Wording

* [ALS-5852] Update footer render function

The update introduces new dependencies in the footer.js file and modifies the render function with additional functionality. A new middleware and redirect modal are instantiated in the function. An event handler is also added for clicks on anchors with target="_blank" to prevent the default behavior and render the redirect modal instead.

* Refactor header.js to remove unnecessary import

In the header.js file, the "header/goToAuthorized" import was removed as it was unnecessary. By doing so, the code becomes cleaner and easier to maintain.

* Improve redirectModal message flexibility

In the redirect-modal.js file, the headerMessage and message have been refactored to accept dynamic inputs from the calling method. Similarly, the footer.js has been modified to pass a custom message to the redirectModal. This enhancement increases flexibility and allows customized messages for different external links.

* [ALS-5885] Update Terms of Use modal header text

* [ALS-5567] Fix not authorized import