Refactor search term selection in search view
The code has been refactored to unify the selection process of search terms in search view. Regardless of open or authorized access, the search term used will now be determined by the `tourSearchTerm` option provided. Default term for open access is "epilepsy" while for authorized access it can vary, for instance, "cardiac surgery" in biodatacatalyst and "race" in aim-ahead.