2024-03-28-Release (#13)
* [ALS-5583] Add Dockerfile for frontend in pic-sure-aim-ahead

Included a biodatacatalyst based Dockerfile in the frontend folder for pic-sure-aim-ahead with specific paths for UI components. The Dockerfile will allow easy deployment and management of the frontend. Optimized building pipeline and prepared for potential UI overwrites later on.

* [ALS-5583] Update Dockerfile source in frontend applications

The Dockerfiles for both 'pic-sure-aim-ahead-frontend' and 'open-pic-sure-bdc-frontend' have been updated to use a new source image. Specifically, it changes from 'biodatacatalyst_ui' to 'hms-dbmi/pic-sure-biodatacatalyst-ui'. This change ensures smoother deployment and improved management of our frontend applications.

* [ALS-5583] Temporarily add Open Access router.js override

I am temporarily adding Open access's router override file, so we can utilize the same login flow for UI development.

* [ALS-5516] Add new logo

* [ALS-5517] Add override header

Removed API and data dashboard.

* [ALS-5476] Add terms of use

* [ALS-5581] Remove footer content with override

* [ALS-5518] Override package-view.hbs

Remove the “Seven Bridges” and “Terra” buttons.

* Initial aim ahead development (#3)

* [ALS-5567] Update displayNotAuthorized function in login.js
The displayNotAuthorized function in login.js has been updated to include a dynamic login link to the notAuthorizedTemplate. This change improves maintainability and decreases potential for errors as URLs won't require hard coding anymore.

* [ALS-5566] Add showLoginPage function in login.js
A new function showLoginPage has been added to login.js that displays a confirmation message upon successful authentication. This is specifically implemented for BioDataCatalyst authentication and the message is shown during the UserProfile information processing phase.

* [GC] Remove unnecessary override files
* [GC] Add logo to images directory in Dockerfile
* [GC] Remove Open Access wording

* [GC] Update header links in interface
The href attribute in the interface header links has been updated to provide proper redirection functionality. This change corrects problems with dead links on the website, specifically in the header section, improving usability and the overall user experience.

* [ALS-5517] Remove unused functions in router.js
The commit removes displayDataAccess, getGenomicFilters, and displayAPI functions from router.js. These functions were not in use, contributing to complexity and potential confusion. The removal of these functions simplifies the codebase and makes maintenance easier.

* Remove Open PIC-SURE link from header
The Open PIC-SURE link in the header navigation has been removed as it was no longer needed. This change simplifies the header component and removes potential confusion for users navigating the UI.

* Add landing page and update header in PIC-SURE UI
* [ALS-5581] Remove footer content with override

* Added a function to generate random state for each authentication request in login.js to mitigate CSRF attacks. Adjusted behavior when a code is present and added checks for state parameter to further secure the login flow. Also improved error message handling for authentication failures.

* Add Super Admin and Admin navigation options
A Super Admin Console drop-down menu and an Admin Users button have been introduced in the header navigation. The Super Admin Console contains links for Application, Connection, Privilege, and Role management. These functionalities are only visible to users with respective Super Admin and Admin privileges.

* [ALS-5779] AIM-AHEAD PIC-SURE: Change header name from Authorized Access to Explore
* [ALS-5779] Redirect to login page if authentication was successful
* [ALS-5781] AIM-AHEAD PIC-SURE: Show the API tab
* [ALS-5800] Set logo as favicon
* [ALS-5781] AIM-AHEAD PIC-SURE: Show the API tab
* [ALS-5567] Fix order and now hide header-tabs during the login process

* Remove console log and inline redirect

* [ALS-5518] Improve styles

* Remove extra spinner

* [ALS-5476] Add header.js override file to fix TOS (#8)

* [ALS-5476] Add new header.js file for custom UI component
This commit adds a new header file in the pic-sure-aim-ahead-frontend project to override the default UI and add custom behavior after the main header has rendered. Specifically, it introduces logic to focus on the help dropdown menu when the terms of service modal is displayed.

* Refactor header.js to remove unnecessary import
In the header.js file, the "header/goToAuthorized" import was removed as it was unnecessary. By doing so, the code becomes cleaner and easier to maintain.

* [ALS-5699] Remove Icons and update sub search text (#7)

* [ALS-5850] Update GitHub Link (#6)

* [ALS-5850] Update logo title to say "AIM-AHEAD" (#5)

* [ALS-5852] AIM-AHEAD PIC-SURE StatViz should work and not show obfuscation (#4)

* [ALS-5852] Update Wording

* [ALS-5852] Update footer render function

The update introduces new dependencies in the footer.js file and modifies the render function with additional functionality. A new middleware and redirect modal are instantiated in the function. An event handler is also added for clicks on anchors with target="_blank" to prevent the default behavior and render the redirect modal instead.

* Minor UI fixes and updates (#9)

* [ALS-5476] Add new header.js file for custom UI component

This commit adds a new header file in the pic-sure-aim-ahead-frontend project to override the default UI and add custom behavior after the main header has rendered. Specifically, it introduces logic to focus on the help dropdown menu when the terms of service modal is displayed.

* [ALS-5699] Remove Icons and update sub search text

* [ALS-5850] Update GitHub Link

* [ALS-5850] Update logo title to say "AIM-AHEAD"

* [ALS-5852] Update Wording

* [ALS-5852] Update footer render function

The update introduces new dependencies in the footer.js file and modifies the render function with additional functionality. A new middleware and redirect modal are instantiated in the function. An event handler is also added for clicks on anchors with target="_blank" to prevent the default behavior and render the redirect modal instead.

* Refactor header.js to remove unnecessary import

In the header.js file, the "header/goToAuthorized" import was removed as it was unnecessary. By doing so, the code becomes cleaner and easier to maintain.

* Improve redirectModal message flexibility

In the redirect-modal.js file, the headerMessage and message have been refactored to accept dynamic inputs from the calling method. Similarly, the footer.js has been modified to pass a custom message to the redirectModal. This enhancement increases flexibility and allows customized messages for different external links.

* [ALS-5885] Update Terms of Use modal header text

* [ALS-5567] Fix not authorized import

* Updated Help Desk and Not Authorized Contact links

* [ALS-5902] Add logout functionality on header (#10)

* Update redirectModal initialization in footer.js

Never passed the headerMessage into AIM footer

* Refactor search term selection in search view (#11)

The code has been refactored to unify the selection process of search terms in search view. Regardless of open or authorized access, the search term used will now be determined by the `tourSearchTerm` option provided. Default term for open access is "epilepsy" while for authorized access it can vary, for instance, "cardiac surgery" in biodatacatalyst and "race" in aim-ahead.